backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
pydantic==2.6.4
python-dotenv==1.0.1
